% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{filter.gene}
\alias{filter.gene}
\title{filter.gene}
\usage{
filter.gene(object, thr)
}
\arguments{
\item{object}{: A SeuratObject.}

\item{thr}{A number.
If thr is between 0 and 1, it indicates the the minimum proportion of expression for each gene across all spots.
If thr is greater than 1, it indicates the minimum number of each gene expressed in the spots.}
}
\value{
A seurat object.
}
\description{
This function filters the non-expressed genes and low expression proportion genes.
Only applicable to SeuratObject type data.
}
