% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{scale.count}
\alias{scale.count}
\title{:scale.count}
\usage{
\method{scale}{count}(data, qh = 0.985)
}
\arguments{
\item{data}{A data.frame or matrix of expression counts (spots x genes)}

\item{qh}{A numeric value should be between 0 and 1.}
}
\value{
A data.frame or matrix of adjusted expression counts (spots x genes)
}
\description{
scale.count, original name: simu.count.
This is a Basic function and will not be called directly by users.
Transformation of expression counts.
To adjust the effects from library size between sample spots and outliers of gene expression.
}
