% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{seurat.trans}
\alias{seurat.trans}
\title{seurat.trans}
\usage{
seurat.trans(object, counts = F, sps = F)
}
\arguments{
\item{object}{:A seurat object.}

\item{counts}{A logic value.T/F
If counts==T, the expression matrix is 'counts' in the 'assays' slots.
If counts==F, the expression matrix is 'data' in the 'assays' slots.}

\item{sps}{A logic value.T/F
If sps==T, the outpot is a sparse matrix.}
}
\value{
A data.frame of matrix of counts and coordinates.
}
\description{
Original name is scale.seu.spk
Transform SeuratObject of ST data to data frame.
}
