% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{simu.scale}
\alias{simu.scale}
\title{simu.scale}
\usage{
simu.scale(data)
}
\arguments{
\item{data}{A data.frame or matrix of expression counts}
}
\value{
A data.frame or matrix.
}
\description{
Basic fucntion."This function will not be called directly by users.
}
