% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{svg.2Dplot}
\alias{svg.2Dplot}
\title{2D plot for genes.}
\usage{
svg.2Dplot(data, gene, method, title = NULL, size = 0.5, log = F)
}
\arguments{
\item{data}{A data frame that contains the data to be visualized.
The data format is the same as the parameter 'data' in heartsvg and svg.clust.
It contains gene coordinates and gene expression counts.}

\item{gene}{A vector, the gene or list of genes to be plotted.}

\item{method}{he data processing method, which can be "raw" or "mean".
When set to "raw", the raw data of the specified genes will be plotted.
When set to "mean", the average expression values of the specified genes
in each sample will be plotted.}

\item{title}{A character string.An optional parameter for the title of the plot.}

\item{size}{point size.}

\item{log}{An optional parameter indicating whether to perform
a log transformation on the data, with a default value of F.
If set to T, the data will be transformed using log1p.}
}
\value{
A 2D plot generated by the package ggplot2.
}
\description{
2D plots for genes.
The color of the data point represents the expression value of
a given gene in each sample.
}
