% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{svg.3Dplot}
\alias{svg.3Dplot}
\title{title  3D plots for genes.}
\usage{
svg.3Dplot(data, gene, type = NULL)
}
\arguments{
\item{data}{A data frame that contains the data to be visualized.
The data format is the same as the parameter 'data' in heartsvg and svg.clust.
It contains gene coordinates and gene expression counts.}

\item{gene}{A vector, the gene or list of genes to be plotted.}

\item{type}{A character string. The type of the plot, can be 'scatter' or 'surface'.
If the type parameter is 'surface', a 3D surface plot will be drawn.
If the type parameter is 'scatter', a 3D scatter plot will be drawn."}
}
\value{
3D plots.
}
\description{
title  3D plots for genes.
}
