% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg_cluster.R
\name{svg.clust}
\alias{svg.clust}
\title{svg.cluster}
\usage{
svg.clust(data, svg, method = "h", n.c = NULL)
}
\arguments{
\item{data}{A data.frame with dimensions n*(p+2) contains gene coordinates and
gene expression counts. The first 2 columns should represent coordinates and
their column names must be 'row' and 'col', respectively.
The remaining p columns correspond to the expression levels of p genes.
There are n rows, representing n spots.
It is the same as the parameter 'data' in the function 'heartsvg()'.}

\item{svg}{A vector,indicating columns(genes) to select from the ST data.}

\item{method}{A character string for clustering method.
'h' represents hierarchical cluster analysis.
'k' represents k-means cluster analysis. If k-means method is chosen,
the parameter 'n.c' have to be provided.}

\item{n.c}{A numerical value of the number of clusters.
For hierarchical clustering, n.c is a dispensable parameter.
If n.c is not provided in hierarchical clustering, our approach adopt
Yamamoto test to estimate the number of SVG cluters.
For k-means method, n.c must be provided.}
}
\value{
A data.frame with 2 columns, the first column ('gene') is the gene name.
The second column ('cluster') is the cluster of each gene.
}
\description{
Predictions of spatial domains based on SVG clustering results.
Cluster analysis on SVGs by the spatial expression similarity.
}
