% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{svg.enrichment}
\alias{svg.enrichment}
\title{Enrichment analysis for SVGs.}
\usage{
svg.enrichment(
  svg,
  plot = T,
  meta = F,
  organism = "hsapiens",
  sources = NULL,
  user_threshold = 0.05,
  correction_method = "g_SCS"
)
}
\arguments{
\item{svg}{A vector or data frame containing gene names for enrichment analysis.}

\item{plot}{A logical value indicating whether to plot the
graphical results of the enrichment analysis. The default is True.}

\item{meta}{A logical value indicating whether to return the complete gprofiler2 results,
including annotations, background genes, etc. The default is False.}

\item{organism}{Organism name. Default value is ‘hspainens'
More details in the R package gprofiler2 or the g:GOSt web tool}

\item{sources}{A vector of data sources to use.
Currently, these include GO (GO:BP, GO:MF, GO:CC to select a particular GO branch),
KEGG, REAC, TF, MIRNA, CORUM, HP, HPA, WP.
Please see the R package gprofiler2 or the g:GOSt web tool for
the comprehensive list and details on incorporated data sources}

\item{user_threshold}{A numeric value.
The p-value threshold for significance. Default value is 0.05.}

\item{correction_method}{The algorithm used for multiple testing correction,
one of "gSCS" (synonyms: "analytical", "g_SCS"),"fdr" (synonyms: "false_discovery_rate"), "bonferroni".}
}
\value{
A data.frame contains the enrichment analysis results.
}
\description{
This function is based on the package gprofiler2
 for gene enrichment analysis.
}
