% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{svg.Mplot}
\alias{svg.Mplot}
\title{marginal plots for genes.}
\usage{
svg.Mplot(data, gene, method = "mean")
}
\arguments{
\item{data}{A data frame that contains the data to be visualized.
The data format is the same as the parameter 'data' in heartsvg and svg.clust.
It contains gene coordinates and gene expression counts.}

\item{gene}{A vector, the gene or list of genes to be plotted.}

\item{method}{A character string for the drawing method,
with optional values of 'mean' or 'raw'.
If set to ‘mean’, it will calculate the average expression value of the
input gene set and plot the corresponding marginal expression plots
in the row and column directions. If set to ‘raw’, it will plot the
marginal expression plots of each gene in the input gene set
in the row and column directions.}
}
\value{
marginal plots along row and column directions for genes.
}
\description{
marginal plots for genes.
}
