% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg_cluster.R
\name{svg.seek}
\alias{svg.seek}
\title{Title svg.seek}
\usage{
svg.seek(data, target, svg, size = NULL)
}
\arguments{
\item{data}{A data.frame with dimensions n*(p+2) contains gene coordinates and
gene expression counts. The first 2 columns should represent coordinates and
their column names must be 'row' and 'col', respectively.}

\item{target}{A character string of the target gene for which you want to find nearby genes.}

\item{svg}{A vector. This is a gene list used to narrow down the search range. 
The function will look for nearby genes within this list.}

\item{size}{This is an optional parameter that specifies how many nearby genes you want to return. 
If provided, the function will return the top size genes closest to the target gene. 
If this parameter is not provided, the function will determine the number of nearby genes
 based on some breakpoint}
}
\description{
Title svg.seek
}
